# OpenWA/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sessions/{sessionId}/labels": {
      "get": {
        "operationId": "LabelController_findAll",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of labels"
          },
          "400": {
            "description": "Session not ready or not a business account"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Get all labels (WhatsApp Business only)",
        "tags": [
          "labels"
        ]
      }
    },
    "/api/sessions/{sessionId}/labels/{labelId}": {
      "get": {
        "operationId": "LabelController_findOne",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labelId",
            "required": true,
            "in": "path",
            "description": "Label ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Label details"
          },
          "404": {
            "description": "Label not found"
          }
        },
        "summary": "Get a specific label by ID",
        "tags": [
          "labels"
        ]
      },
      "put": {
        "description": "Baileys only. `PUT` rather than `POST` because the label id is chosen by the caller: WhatsApp carries one `label_edit` write keyed on that id, so whether this creates or updates depends purely on whether the id already exists, and there is no server-assigned id to return.\n\n**Choose an unused id to create.** Reusing one silently rewrites that label rather than failing, because the protocol has no create-only form. Fields left out are left alone.\n\nwhatsapp-web.js can read and assign labels but cannot edit one, and answers `501`.",
        "operationId": "LabelController_upsertLabel",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labelId",
            "required": true,
            "in": "path",
            "description": "Label ID — caller-chosen",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertLabelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Label created or updated"
          },
          "400": {
            "description": "Session not started, or validation failed"
          },
          "501": {
            "description": "The active engine cannot edit labels (whatsapp-web.js)"
          }
        },
        "summary": "Create or update a label",
        "tags": [
          "labels"
        ]
      },
      "delete": {
        "description": "Baileys only. The label disappears from every chat it was on. whatsapp-web.js cannot edit labels and answers `501`.",
        "operationId": "LabelController_deleteLabel",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labelId",
            "required": true,
            "in": "path",
            "description": "Label ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Label deleted"
          },
          "400": {
            "description": "Session not started"
          },
          "501": {
            "description": "The active engine cannot edit labels (whatsapp-web.js)"
          }
        },
        "summary": "Delete a label",
        "tags": [
          "labels"
        ]
      }
    },
    "/api/sessions/{sessionId}/labels/{labelId}/chats": {
      "get": {
        "description": "whatsapp-web.js only. Baileys exposes label writes but no label query of any kind, so it answers `501`.",
        "operationId": "LabelController_getChatsByLabel",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labelId",
            "required": true,
            "in": "path",
            "description": "Label ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chats carrying the label"
          },
          "400": {
            "description": "Session not started"
          },
          "501": {
            "description": "The active engine cannot list chats by label (Baileys)"
          }
        },
        "summary": "Get every chat carrying a label",
        "tags": [
          "labels"
        ]
      }
    },
    "/api/sessions/{sessionId}/labels/chat/{chatId}": {
      "get": {
        "operationId": "LabelController_getChatLabels",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "Chat ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of labels for the chat"
          }
        },
        "summary": "Get labels for a specific chat",
        "tags": [
          "labels"
        ]
      },
      "post": {
        "operationId": "LabelController_addLabelToChat",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "Chat ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "labelId": {
                    "type": "string",
                    "description": "Label ID to add"
                  }
                },
                "required": [
                  "labelId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Label added to chat"
          },
          "422": {
            "description": "Labels require a WhatsApp Business account, or the chat type has no labels"
          }
        },
        "summary": "Add a label to a chat",
        "tags": [
          "labels"
        ]
      }
    },
    "/api/sessions/{sessionId}/labels/chat/{chatId}/{labelId}": {
      "delete": {
        "operationId": "LabelController_removeLabelFromChat",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "Chat ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labelId",
            "required": true,
            "in": "path",
            "description": "Label ID to remove",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Label removed from chat"
          },
          "422": {
            "description": "Labels require a WhatsApp Business account, or the chat type has no labels"
          }
        },
        "summary": "Remove a label from a chat",
        "tags": [
          "labels"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.13.0",
    "contact": {
      "name": "OpenWA",
      "url": "https://github.com/rmyndharis/OpenWA",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [
    {
      "url": "http://{host}:{port}",
      "description": "OpenWA instance",
      "variables": {
        "host": {
          "default": "localhost"
        },
        "port": {
          "default": "2785",
          "description": "PORT env var"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "metrics-bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "METRICS_TOKEN for the GET /api/metrics scrape endpoint"
      }
    },
    "schemas": {
...
      "UpsertLabelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Label name. Omit to leave the current name untouched.",
            "example": "VIP customer",
            "minLength": 1,
            "maxLength": 100
          },
          "color": {
            "type": "number",
            "description": "WhatsApp's colour INDEX (0-19), not a hex value — the read path's `hexColor` cannot be translated back, because neither engine exposes the mapping. Omit to leave the colour alone.",
            "example": 3,
            "minimum": 0,
            "maximum": 19
          }
        }
      },
...
    }
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}